Refactor common methods in http server service and indexing server service
