+++ packages/core/package.json
@@ -60,7 +60,6 @@
     "express": "^4.18.1",
     "glob": "^8.1.0",
     "graphql": "^15.3.0",
-    "graphql-http": "^1.22.0",
     "graphql-subscriptions": "^2.0.0",
     "graphql-ws": "^5.14.0",
     "http-terminator": "^3.2.0",
