+++ packages/core/src/server/service.ts
@@ -1,37 +1,39 @@
-import cors from "cors";
-import express from "express";
 import type { FormattedExecutionResult, GraphQLSchema } from "graphql";
 import { formatError, GraphQLError } from "graphql";
 import { createHandler } from "graphql-http/lib/use/express";
-import { createHttpTerminator } from "http-terminator";
-import { createServer, Server } from "node:http";
 
 import type { Common } from "@/Ponder.js";
 import { graphiQLHtml } from "@/ui/graphiql.html.js";
 import type { UserStore } from "@/user-store/store.js";
+import { Server } from "@/utils/server.js";
 import { startClock } from "@/utils/timer.js";
 
 export class ServerService {
   private common: Common;
   private userStore: UserStore;
-
-  private port: number;
-  app?: express.Express;
-
-  private terminate?: () => Promise<void>;
+  private server: Server;
 
   isHistoricalIndexingComplete = false;
 
   constructor({ common, userStore }: { common: Common; userStore: UserStore }) {
     this.common = common;
     this.userStore = userStore;
-    this.port = this.common.options.port;
+
+    this.server = new Server({
+      common,
+      port: common.options.port,
+    });
+  }
+
+  get app() {
+    return this.server.app;
   }
 
   async start() {
-    this.app = express();
-    this.app.use(cors({ methods: ["GET", "POST", "OPTIONS", "HEAD"] }));
-    this.app.use((req, res, next) => {
+    await this.server.start();
+    this.common.metrics.ponder_server_port.set(this.server.port);
+
+    this.server.app?.use((req, res, next) => {
       const endClock = startClock();
       res.on("finish", () => {
         const responseDuration = endClock();
@@ -64,79 +66,6 @@ export class ServerService {
         );
       });
       next();
-    });
-
-    const server = await new Promise<Server>((resolve, reject) => {
-      const server = createServer(this.app)
-        .on("error", (error) => {
-          if ((error as any).code === "EADDRINUSE") {
-            this.common.logger.warn({
-              service: "server",
-              msg: `Port ${this.port} was in use, trying port ${this.port + 1}`,
-            });
-            this.port += 1;
-            setTimeout(() => {
-              server.close();
-              server.listen(this.port);
-            }, 5);
-          } else {
-            reject(error);
-          }
-        })
-        .on("listening", () => {
-          this.common.metrics.ponder_server_port.set(this.port);
-          resolve(server);
-        })
-        .listen(this.port);
-    });
-
-    const terminator = createHttpTerminator({ server });
-    this.terminate = () => terminator.terminate();
-
-    this.common.logger.info({
-      service: "server",
-      msg: `Started listening on port ${this.port}`,
-    });
-
-    this.app.post("/metrics", async (_, res) => {
-      try {
-        res.set("Content-Type", "text/plain; version=0.0.4; charset=utf-8");
-        res.end(await this.common.metrics.getMetrics());
-      } catch (error) {
-        res.status(500).end(error);
-      }
-    });
-
-    this.app.get("/metrics", async (_, res) => {
-      try {
-        res.set("Content-Type", "text/plain; version=0.0.4; charset=utf-8");
-        res.end(await this.common.metrics.getMetrics());
-      } catch (error) {
-        res.status(500).end(error);
-      }
-    });
-
-    // By default, the server will respond as unhealthy until historical index has
-    // been processed OR 4.5 minutes have passed since the app was created. This
-    // enables zero-downtime deployments on PaaS platforms like Railway and Render.
-    // Also see https://github.com/0xOlias/ponder/issues/24
-    this.app.get("/health", (_, res) => {
-      if (this.isHistoricalIndexingComplete) {
-        return res.status(200).send();
-      }
-
-      const max = this.common.options.maxHealthcheckDuration;
-      const elapsed = Math.floor(process.uptime());
-
-      if (elapsed > max) {
-        this.common.logger.warn({
-          service: "server",
-          msg: `Historical sync duration has exceeded the max healthcheck duration of ${max} seconds (current: ${elapsed}). Sevice is now responding as healthy and may serve incomplete data.`,
-        });
-        return res.status(200).send();
-      }
-
-      return res.status(503).send();
     });
   }
 
@@ -213,15 +142,12 @@ export class ServerService {
   }
 
   async kill() {
-    await this.terminate?.();
-    this.common.logger.debug({
-      service: "server",
-      msg: `Stopped listening on port ${this.port}`,
-    });
+    await this.server.kill();
   }
 
-  setIsHistoricalIndexingComplete() {
+  setIsHistoricalEventProcessingComplete() {
     this.isHistoricalIndexingComplete = true;
+    this.server.isHealthy = true;
 
     this.common.logger.info({
       service: "server",
